modify configs.

+++ _hexo/scripts/clean.js
@@ -1,3 +1,6 @@
 var fs = require("fs");
 
-fs.unlink(__dirname+"/../db.json");
+if (process.argv[2] == "server") {
+	console.log("remove db.json");
+	fs.unlink(__dirname+"/../db.json");
+}
